+++ src/services/CertificatePayloadGenerator.ts
@@ -1,7 +1,7 @@
 import { Inject, Service } from 'typedi';
 import { CERTIFICATE_DATA } from '../models/Enums';
 import { ITestResult } from '../models/ITestResult';
-import { CertificatePayloadGeneratorIva } from './CertificatePayloadGeneratorIva';
+import { CertificatePayloadGeneratorIva } from './certificate-payload/CertificatePayloadGeneratorIva';
 import { CertificatePayloadGeneratorAdr } from './CertificatePayloadGeneratorAdr';
 import { CertificatePayloadGeneratorRwt } from './CertificatePayloadGeneratorRwt';
 import { CertificatePayloadGeneratorPassOrFail } from './CertificatePayloadGeneratorPassOrFail';
